+++ tests/testthat/test-old-school.R
@@ -1,8 +1,13 @@
 context("old-school expectations")
 
 test_that("old school logical works", {
-  expect_success(expect_that(TRUE, is_true()))
-  expect_success(expect_that(FALSE, is_false()))
+  expect_warning(
+    expect_success(expect_that(TRUE, is_true())),
+    "deprecated")
+
+  expect_warning(
+    expect_success(expect_that(FALSE, is_false())),
+    "deprecated")
 })
 
 test_that("old school types still work", {